deploy fragmentation complete
